move to test_models.py and use default logic there
